Add dotenv for easier local development
This allows running the webpack dev server without env variables on your
command line, so you don't have to copy/paste the `npm start` line or
search for it in your terminal's history.

The `.env` loading is ignored if there is no `.env` file, so previous
use-cases + non-local environments continue to work the same way as before.